fix DealVote and ReportVote plate

+++ selab-vote/src/main/java/tyut/selab/vote/service/impl/DealVoteServiceImpl.java
@@ -12,24 +12,20 @@
  * <desc> 版本描述
  */
 package tyut.selab.vote.service.impl;
-
-
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tyut.selab.vote.domain.DTO.VoteInfoLaunchDTO;
-import tyut.selab.vote.domain.po.VoteInfo;
-import tyut.selab.vote.domain.po.VoteOption;
 import tyut.selab.vote.enums.VoteStatus;
+import tyut.selab.vote.exception.VoteDeletedException;
+import tyut.selab.vote.exception.VoteFreezedException;
+import tyut.selab.vote.exception.VoteOverTimeException;
+import tyut.selab.vote.exception.VoteProcessedException;
+import tyut.selab.vote.mapper.ReportVoteMapper;
 import tyut.selab.vote.mapper.VoteInfoMapper;
 import tyut.selab.vote.mapper.VoteOptionMapper;
 import tyut.selab.vote.mapper.VoteWeightMapper;
 import tyut.selab.vote.service.DealVoteService;
 import tyut.selab.vote.tools.TimeDealTool;
-
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 
 /**
  * @ClassName: DealVoteServiceImpl
@@ -49,56 +45,62 @@ public class DealVoteServiceImpl implements DealVoteService {
     @Autowired
     private VoteOptionMapper voteOptionMapper;
 
+    @Autowired
+    private ReportVoteMapper reportVoteMapper;
 
     @Override
-    public Integer launchVote(VoteInfoLaunchDTO voteInfoLaunchDTO) {
+    public void launchVote(VoteInfoLaunchDTO voteInfoLaunchDTO) {
 //        Long userId = SecurityUtils.getUserId();
-//        voteInfo.setUserId(userId);
-
+//       voteInfo.setUserId(userId);
         voteInfoLaunchDTO.setStatus(VoteStatus.UNDERWAY);
-        // TODO 是否需要判断截止时间合理性
+        voteInfoLaunchDTO.setDelFlag("1");
         voteInfoMapper.saveVoteInformation(voteInfoLaunchDTO);
-        List<VoteOption> voteOptionList = new ArrayList<>();
-        voteInfoLaunchDTO.getVoteOptionVoList().forEach(voteOptionVo -> {
-            VoteOption voteOption = new VoteOption();
-            voteOption.setVoteId(voteOptionVo.getVoteId());
-            voteOption.setOptionType(voteOptionVo.getOptionType());
-            voteOption.setContent(voteOptionVo.getContent());
-            voteOptionList.add(voteOption);
-        });
-        voteOptionMapper.saveVoteOptionInformation(voteOptionList);
+        voteOptionMapper.saveVoteOptionInformation(voteInfoLaunchDTO.getVoteOptionVoList());
         voteWeightMapper.saveVoteWeightInformation(voteInfoLaunchDTO.getVoteWeights());
-        return null;
     }
 
     @Override
-    public Integer withdrawVote(Long voteId) {
-        return null;
-    }
-
-
-
-    @Override
-    public Integer HandlingFrozenVote(Long voteId,Integer handel) {
+    public void withdrawVote(Long voteId) throws VoteOverTimeException, VoteFreezedException {
         if(TimeDealTool.judgeVoteFinish(voteInfoMapper.queryVoteDeadTime(voteId))){
-            //未到截止时间
-            //handel为1则恢复正常
-            if(handel == 1){
-                voteInfoMapper.updateVoteStatus(voteId, VoteStatus.UNDERWAY);
-                return 2;
-            }else{
-                voteInfoMapper.updateVoteStatus(voteId, VoteStatus.CLOSED);
-                return 1;
+            //未到截止时间,撤回投票
+            //判断一下是否已被冻结
+            if(voteInfoMapper.getVoteStatus(voteId).getStatus() == VoteStatus.FREEZE){
+                throw new VoteFreezedException("该投票已被冻结");
             }
+            voteInfoMapper.updateVoteStatus(voteId, VoteStatus.WITHDRAW);
         }else{
-            //此刻超出截止时间，结束投票
-            voteInfoMapper.updateVoteStatus(voteId,VoteStatus.FINISH);
+            throw new VoteOverTimeException("该投票已结束");
+        }
+    }
+
+    @Override
+    public Integer HandlingFrozenVote(Long voteId,Integer handel) throws VoteOverTimeException, VoteProcessedException {
+        if(!TimeDealTool.judgeVoteFinish(voteInfoMapper.queryVoteDeadTime(voteId)))
+            throw new VoteOverTimeException("该投票已结束");
+
+        VoteStatus voteStatus = voteInfoMapper.getVoteStatus(voteId).getStatus();
+        if(voteStatus == VoteStatus.CLOSED || voteStatus == VoteStatus.UNDERWAY)
+            throw new VoteProcessedException("该投票已被处理");
+
+        //handel为1则恢复正常
+        if(handel == 1){
+            if(voteStatus == VoteStatus.FREEZE){
+                //解除冻结后举报数要清0
+                reportVoteMapper.clearReportCount(voteId);
+            }
+            voteInfoMapper.updateVoteStatus(voteId, VoteStatus.UNDERWAY);
+            return 1;
+        }else{
+            voteInfoMapper.updateVoteStatus(voteId, VoteStatus.CLOSED);
             return 0;
         }
     }
 
     @Override
-    public Integer deleteVote(Long voteId) {
+    public Integer deleteVote(Long voteId) throws VoteDeletedException {
+        //判断是否已被删除
+        if(voteInfoMapper.getVoteDelFlag(voteId) == 0)
+            throw new VoteDeletedException("该投票已被删除");
 
         return voteInfoMapper.deleteVote(voteId);
     }
